Fix bug where assigning to errno in assembly would overwrite 4 bytes of the program break, breaking future malloc calls which were large enough

+++ src/libc/stdlib.c
@@ -6,7 +6,7 @@
  * Last updated 2021-11-15
  */
 
-int errno;
+int64_t errno;
 
 void abort(void)
 {
